+++ server/src/main/java/org/allaymc/server/block/component/BlockNoteblockBaseComponentImpl.java
@@ -2,18 +2,20 @@ package org.allaymc.server.block.component;
 
 import org.allaymc.api.block.BlockBehavior;
 import org.allaymc.api.block.data.BlockFace;
-import org.allaymc.api.block.data.BlockTags;
 import org.allaymc.api.block.data.Instrument;
 import org.allaymc.api.block.dto.Block;
+import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.type.BlockType;
 import org.allaymc.api.block.type.BlockTypes;
 import org.allaymc.api.blockentity.interfaces.BlockEntityNoteblock;
+import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.world.Dimension;
 import org.allaymc.api.world.particle.NoteParticle;
 import org.allaymc.api.world.sound.NoteSound;
 
 /**
  * Implementation of the noteblock.
- * Plays sound when powered by redstone (on rising edge).
+ * Plays sound when powered by redstone (on rising edge) or when interacted.
  *
  * @author daoge_cmd
  */
@@ -40,6 +42,37 @@ public class BlockNoteblockBaseComponentImpl extends BlockBaseComponentImpl {
         blockEntity.setPowered(powered);
     }
 
+    @Override
+    public boolean onInteract(ItemStack itemStack, Dimension dimension, PlayerInteractInfo interactInfo) {
+        if (interactInfo == null) {
+            return false;
+        }
+
+        // Sneaking player should not trigger noteblock (allows placing blocks on top)
+        if (interactInfo.player().isSneaking()) {
+            return false;
+        }
+
+        var clickedBlock = interactInfo.getClickedBlock();
+        // Can't play if block above is not air
+        if (clickedBlock.offsetPos(BlockFace.UP).getBlockType() != BlockTypes.AIR) {
+            return false;
+        }
+
+        var blockEntity = (BlockEntityNoteblock) clickedBlock.getBlockEntity();
+        if (blockEntity == null) {
+            return false;
+        }
+
+        var instrument = Instrument.fromBlockBelow(clickedBlock.offsetPos(BlockFace.DOWN).getBlockType());
+        var pitch = blockEntity.getNextPitch();
+
+        clickedBlock.addSound(new NoteSound(instrument, pitch));
+        // NOTICE: Do not use Block.addParticle() method which center the block pos but the client doesn't want
+        clickedBlock.getDimension().addParticle(clickedBlock.getPosition(), new NoteParticle(instrument, pitch));
+        return true;
+    }
+
     /**
      * Emits sound and particle from the noteblock.
      */
@@ -54,80 +87,11 @@ public class BlockNoteblockBaseComponentImpl extends BlockBaseComponentImpl {
             return;
         }
 
-        var instrument = getInstrument(block.offsetPos(BlockFace.DOWN).getBlockType());
+        var instrument = Instrument.fromBlockBelow(block.offsetPos(BlockFace.DOWN).getBlockType());
         var pitch = blockEntity.getPitch();
 
         block.addSound(new NoteSound(instrument, pitch));
-        block.addParticle(new NoteParticle(instrument, pitch));
-    }
-
-    /**
-     * Gets the instrument based on the block below the noteblock.
-     */
-    protected Instrument getInstrument(BlockType<?> block) {
-        if (block.hasBlockTag(BlockTags.WOOD) ||
-            block.hasBlockTag(BlockTags.LOG)) {
-            return Instrument.BASS;
-        }
-
-        if (block.hasBlockTag(BlockTags.SAND) ||
-            block.hasBlockTag(BlockTags.GRAVEL) ||
-            block.hasBlockTag(BlockTags.CONCRETE_POWDER) ||
-            block == BlockTypes.HEAVY_CORE) {
-            return Instrument.SNARE;
-        }
-
-        if (block.hasBlockTag(BlockTags.STONE)) {
-            return Instrument.BASS_DRUM;
-        }
-
-        if (block == BlockTypes.GOLD_BLOCK) {
-            return Instrument.BELL;
-        }
-
-        if (block.hasBlockTag(BlockTags.INFESTED) ||
-            block == BlockTypes.CLAY ||
-            block == BlockTypes.HONEYCOMB_BLOCK ||
-            block == BlockTypes.RESIN_BLOCK) {
-            return Instrument.FLUTE;
-        }
-
-        if (block == BlockTypes.PACKED_ICE) {
-            return Instrument.CHIMES;
-        }
-
-        if (block.hasBlockTag(BlockTags.WOOL)) {
-            return Instrument.GUITAR;
-        }
-
-        if (block == BlockTypes.BONE_BLOCK) {
-            return Instrument.XYLOPHONE;
-        }
-
-        if (block == BlockTypes.IRON_BLOCK) {
-            return Instrument.IRON_XYLOPHONE;
-        }
-
-        if (block == BlockTypes.SOUL_SAND) {
-            return Instrument.COW_BELL;
-        }
-
-        if (block == BlockTypes.PUMPKIN) {
-            return Instrument.DIDGERIDOO;
-        }
-
-        if (block == BlockTypes.EMERALD_BLOCK) {
-            return Instrument.BIT;
-        }
-
-        if (block == BlockTypes.HAY_BLOCK) {
-            return Instrument.BANJO;
-        }
-
-        if (block == BlockTypes.GLOWSTONE) {
-            return Instrument.PLING;
-        }
-
-        return Instrument.PIANO;
+        // NOTICE: Do not use Block.addParticle() method which center the block pos but the client doesn't want
+        block.getDimension().addParticle(block.getPosition(), new NoteParticle(instrument, pitch));
     }
 }